+++ src/app/dimApp.module.ts
@@ -2,9 +2,7 @@ import { module } from 'angular';
 
 import AriaModule from 'angular-aria';
 import 'ng-dialog';
-import DragAndDropModule from 'angular-native-dragdrop';
 
-import 'angularjs-slider';
 import ToasterModule from 'angularjs-toaster';
 import ngI18Next from 'ng-i18next';
 import 'angular-hotkeys';
@@ -14,7 +12,6 @@ import ngimportMoreModule from './ngimport-more';
 import ngSanitize from 'angular-sanitize';
 
 import { ShellModule } from './shell/shell.module';
-import inventoryModule from './inventory/inventory.module';
 import itemReviewModule from './item-review/item-review.module';
 import infuseModule from './infuse/infuse.module';
 import farmingModule from './farming/farming.module';
@@ -25,14 +22,11 @@ import run from './dimApp.run';
 
 const dependencies = [
   AriaModule,
-  DragAndDropModule,
   ngI18Next,
   ngSanitize,
   ocLazyLoadModule,
   ShellModule,
-  'rzModule',
   ToasterModule,
-  inventoryModule,
   itemReviewModule,
   infuseModule,
   farmingModule,