+++ src/app/inventory/dim-item-info.ts
@@ -148,23 +148,3 @@ function getInfos(key: string): Promise<{ [itemInstanceId: string]: DimItemInfo 
 function setInfos(key: string, infos: { [itemInstanceId: string]: DimItemInfo }) {
   return SyncService.set({ [key]: infos });
 }
-
-export function tagIconFilter() {
-  'ngInject';
-  const iconType: { [P in TagValue]?: IconDefinition | undefined } = {};
-
-  itemTags.forEach((tag) => {
-    if (tag.type) {
-      iconType[tag.type] = tag.icon;
-    }
-  });
-
-  return function tagIcon(value: TagValue) {
-    const icon = iconType[value];
-    if (icon) {
-      return `item-tag fa fa-${icon.iconName}`;
-    } else {
-      return 'item-tag no-tag';
-    }
-  };
-}