+++ src/app/settings/reducer.ts
@@ -57,7 +57,7 @@ export interface Settings {
 }
 
 export function defaultItemSize() {
-  return 48;
+  return 50;
 }
 
 export const initialState: Settings = {