+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove the ability to set a specific vault width. Vault always takes all remaining space.
 * Inventory columns are shaded to match the equipped emblem.
 * Fit and finish changes to the new tiles and inventory display.
+* Add id column to exported csv for ghosts, armor, and weapons
 
 # 5.2.1 (2018-11-20)